requirements.txt
Flask==2.0.2
line-bot-sdk==2.1.0